standardise and refactor test data to be available to all test suites
