+++ src/tests/unit/services/PostEventProcessor.test.ts
@@ -11,7 +11,7 @@ import {
 	VALID_AUTH_IPV_AUTHORISATION_REQUESTED_TXMA_EVENT_STRING,
 	VALID_F2F_YOTI_START_TXMA_EVENT_STRING,
 	VALID_IPV_F2F_CRI_VC_CONSUMED_TXMA_EVENT_STRING,
-} from "../data/sqs-events";
+} from "../../data/sqs-events";
 
 let postEventProcessor: PostEventProcessor;
 const mockIprService = mock<IPRService>();
